refactor(relayer): split sidecar-server.ts monolith into modules (WALM-105)
Restructure the 2,400-line scripts/sidecar-server.ts into focused modules
under scripts/sidecar/ for maintainability. No behavior change: route
paths, auth/middleware order, request/response shapes, log strings and
structured log events are preserved.

New layout:
- config.ts          env-driven configuration (parsed once at load)
- util.ts / log.ts   error/log/request-id helpers
- state.ts           metrics counters + state snapshot
- clients.ts         Sui/SEAL/Walrus clients, refresh lifecycle, tip cache
- concurrency.ts     AsyncSemaphore + global/per-wallet upload limiting
- enoki.ts           Enoki API client + sponsored execution
- wallet.ts          wallet tx submission (metrics, rebuildable retry)
- blob-metadata.ts   metadata-stamp + transfer-to-owner transaction
- retry/rpc.ts       generic RPC retry with backoff + jitter
- routes/            one module per endpoint group
- app.ts / server.ts express assembly + bootstrap
sidecar-server.ts stays as the entry shim (main.rs spawns it by name).

Intentional cleanups while moving:
- seal_approve PTB building deduplicated between decrypt and decrypt-batch
- getDynamicFieldWithRetry folded into withRpcRetry (identical backoff)
- dead "req.body.blobs = ..." write removed from legacy /walrus/set-metadata

Also: Dockerfile copies scripts/sidecar/, mcp_proxy.rs doc comment points
to the new mount location, and unit tests added for the extracted
semaphore, util and query helpers (78 tests pass).

Linear: WALM-105
